+++ spec/dummy_app/app/models/matrix_pill.rb
@@ -2,4 +2,5 @@
 
 # An STI model
 class MatrixPill < ApplicationRecord
+  include IronTrail::Model
 end