+++ spec/models/guitar_spec.rb
@@ -3,6 +3,149 @@
 RSpec.describe Guitar do
   let(:person) { Person.create!(first_name: 'Arthur', last_name: 'Schopenhauer') }
 
+  describe 'IronTrail::ChangeModelConcern' do
+    describe 'helper methods' do
+      before do
+        person.update!(first_name: 'Joe')
+        person.update!(first_name: 'Joey')
+        person.destroy!
+      end
+
+      it 'correctly classifies operations' do
+        expect(person.iron_trails.length).to eq(4)
+        expect(person.iron_trails.inserts.length).to eq(1)
+        expect(person.iron_trails.updates.length).to eq(2)
+        expect(person.iron_trails.deletes.length).to eq(1)
+
+        expect(person.iron_trails.inserts[0].insert_operation?).to be(true)
+        expect(person.iron_trails.inserts[0].update_operation?).to be(false)
+        expect(person.iron_trails.inserts[0].delete_operation?).to be(false)
+
+        expect(person.iron_trails.updates[0].insert_operation?).to be(false)
+        expect(person.iron_trails.updates[0].update_operation?).to be(true)
+        expect(person.iron_trails.updates[0].delete_operation?).to be(false)
+        expect(person.iron_trails.updates[1].insert_operation?).to be(false)
+        expect(person.iron_trails.updates[1].update_operation?).to be(true)
+        expect(person.iron_trails.updates[1].delete_operation?).to be(false)
+
+        expect(person.iron_trails.deletes[0].insert_operation?).to be(false)
+        expect(person.iron_trails.deletes[0].update_operation?).to be(false)
+        expect(person.iron_trails.deletes[0].delete_operation?).to be(true)
+      end
+    end
+  end
+
+  describe 'iron_trails.travel_to' do
+    let(:guitar) { Guitar.create!(description: 'the guitar', person:) }
+
+    before do
+      fake_timestamps = [
+        Time.parse('2005-04-15T13:44:59Z'),
+        Time.parse('2005-05-08T14:00:03Z'),
+        Time.parse('2005-05-08T14:00:15Z'),
+        Time.parse('2005-05-08T14:02:00Z')
+      ]
+
+      guitar.update!(description: 'guitar 2')
+      guitar.update!(description: 'guitar 3')
+      guitar.update!(description: 'guitar 4')
+      @trail_ids = guitar.iron_trails.order(id: :asc).pluck(:id)
+
+      expect(@trail_ids.length).to eq(4)
+
+      @trail_ids.zip(fake_timestamps).each do |trail_id, fake_ts|
+        query = "UPDATE irontrail_changes SET created_at='#{fake_ts}' WHERE id=#{trail_id}"
+        result = ActiveRecord::Base.connection.execute(query)
+        expect(result.cmd_tuples).to eq(1)
+      end
+      guitar.reload
+    end
+
+    it 'recovers the correct record' do
+      git = guitar.iron_trails.travel_to('2005-05-08T14:00:03Z')
+
+      expect(git).to be_a(Guitar)
+      expect(git.id).to eq(guitar.id)
+      expect(git.description).to eq('guitar 2')
+    end
+
+    it 'has no ghost reified attributes' do
+      git = guitar.iron_trails.travel_to('2005-05-08T14:00:03Z')
+      expect(git.irontrail_reified_ghost_attributes).to be_nil
+    end
+
+    context 'when a version has attributes that dont exist anymore' do
+      before do
+        trail_id = @trail_ids[2]
+        trail = IrontrailChange.find_by!(id: trail_id)
+
+        rec_old = trail.rec_old.merge('foo' => 'perfectly fine')
+        rec_new = trail.rec_new.merge('foo' => 'ghosted!')
+
+        query = <<~SQL
+          UPDATE irontrail_changes SET
+            rec_old=#{ActiveRecord::Base.connection.quote(JSON.dump(rec_old))}::jsonb,
+            rec_new=#{ActiveRecord::Base.connection.quote(JSON.dump(rec_new))}::jsonb
+          WHERE id=#{trail_id}
+        SQL
+
+        result = ActiveRecord::Base.connection.execute(query)
+        expect(result.cmd_tuples).to eq(1)
+      end
+
+      describe 'on time' do
+        let(:git) { guitar.iron_trails.travel_to('2005-05-08T14:00:15Z') }
+
+        it 'contains ghost reified attributes' do
+          expect(git).to be_a(Guitar)
+          expect(git.id).to eq(guitar.id)
+          expect(git.description).to eq('guitar 3')
+          expect(git.irontrail_reified_ghost_attributes).to eq({ foo: 'ghosted!' }.with_indifferent_access)
+        end
+      end
+
+      describe 'a little late' do
+        let(:git) { guitar.iron_trails.travel_to('2005-05-08T14:00:17Z') }
+
+        it 'contains ghost reified attributes' do
+          expect(git).to be_a(Guitar)
+          expect(git.description).to eq('guitar 3')
+          expect(git.irontrail_reified_ghost_attributes).to eq({ foo: 'ghosted!' }.with_indifferent_access)
+        end
+      end
+    end
+
+    context 'when the object has been destroyed' do
+      let(:destroy_time) { '2006-10-21T06:00:00Z' }
+      before do
+        guitar.destroy!
+        query = "UPDATE irontrail_changes SET created_at='#{destroy_time}' WHERE operation='d' AND rec_id='#{guitar.id}'"
+        result = ActiveRecord::Base.connection.execute(query)
+        expect(result.cmd_tuples).to eq(1)
+      end
+
+      describe 'on time' do
+        let(:git) { guitar.iron_trails.travel_to(destroy_time) }
+
+        it 'recovers the correct record' do
+          expect(git).to be_a(Guitar)
+          expect(git.id).to eq(guitar.id)
+          expect(git.description).to eq('guitar 4')
+        end
+      end
+
+      describe 'a little late' do
+        let(:git) { guitar.iron_trails.travel_to(Time.parse(destroy_time) + 5) }
+
+        it 'recovers the correct record' do
+          expect(git).to be_a(Guitar)
+          expect(git.id).to eq(guitar.id)
+          expect(git.description).to eq('guitar 4')
+        end
+      end
+    end
+  end
+
   describe 'has_many trails' do
     it 'has the trails' do
       classics = PeopleManager::CLASSIC_GUITARS